+++ Kernel/System/ProcessManagement/TransitionAction/TicketCreate.pm
@@ -25,6 +25,7 @@ our @ObjectDependencies = (
     'Kernel::System::State',
     'Kernel::System::Ticket',
     'Kernel::System::Time',
+    'Kernel::System::User',
 );
 
 =head1 NAME
@@ -81,7 +82,7 @@ sub new {
             State         => 'new',              # or StateID => 5,
             CustomerID    => '123465',
             CustomerUser  => 'customer@example.com',
-            OwnerID       => 123,
+            Owner         => 'someuserlogin',    # or OwnerID => 123
 
             # ticket optional:
             TN              => $TicketObject->TicketCreateNumber(), # optional
@@ -194,6 +195,13 @@ sub Run {
         if ( !$TicketParam{$Attribute} && !$TicketParam{ $Attribute . "ID" } ) {
             $TicketParam{$Attribute} = $Kernel::OM->Get('Kernel::Config')->Get("Process::Default$Attribute") || '';
         }
+    }
+
+    # Get OwnerID from Owner
+    if ( $TicketParam{Owner} && !$TicketParam{OwnerID} ) {
+        $TicketParam{OwnerID} = $Kernel::OM->Get('Kernel::System::User')->UserLookup(
+            UserLogin => $TicketParam{Owner},
+        );
     }
 
     # get ticket object